requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyMuPDF==1.25.0
google-genai==1.53.0
python-dotenv==1.0.0
pytz==2024.1